Add stdout module and WLF
